# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartReserve — Smart Demand-Aware Offer Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
}

export interface AuthResponse {
  token: string;
  user: AdminUser;
}

export interface Business {
  id: number;
  name: string;
  businessType: string;
  ownerName: string;
  phone: string;
  email: string;
  address: string;
  city: string;
  openingTime: string;
  closingTime: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  createdAt?: string;
}

export interface BusinessInput {
  name: string;
  businessType: string;
  ownerName: string;
  phone: string;
  email: string;
  address: string;
  city: string;
  openingTime: string;
  closingTime: string;
  logoUrl?: string;
  bannerUrl?: string;
}

export interface BusinessUpdate {
  name?: string;
  businessType?: string;
  ownerName?: string;
  phone?: string;
  email?: string;
  address?: string;
  city?: string;
  openingTime?: string;
  closingTime?: string;
  logoUrl?: string;
  bannerUrl?: string;
}

export interface Offer {
  id: number;
  businessId: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  discountPercent: number;
  startDate: string;
  endDate: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  terms?: string | null;
  status: string;
  /** @nullable */
  bannerUrl?: string | null;
  createdAt?: string;
}

export interface OfferInput {
  businessId: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  startDate: string;
  endDate: string;
  startTime: string;
  endTime: string;
  terms?: string;
  status?: string;
  bannerUrl?: string;
}

export interface OfferUpdate {
  title?: string;
  description?: string;
  category?: string;
  originalPrice?: number;
  offerPrice?: number;
  startDate?: string;
  endDate?: string;
  startTime?: string;
  endTime?: string;
  terms?: string;
  status?: string;
  bannerUrl?: string;
}

export interface OfferStatusUpdate {
  status: string;
}

export interface Slot {
  id: number;
  offerId: number;
  slotDate: string;
  startTime: string;
  endTime: string;
  capacity: number;
  bookedCount: number;
  availableCount: number;
  status: string;
  occupancyPercent?: number;
  demandLevel?: string;
}

export interface OfferDetail {
  id: number;
  businessId: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  discountPercent: number;
  startDate: string;
  endDate: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  terms?: string | null;
  status: string;
  /** @nullable */
  bannerUrl?: string | null;
  business: Business;
  slots: Slot[];
  createdAt?: string;
}

export interface SlotInput {
  slotDate: string;
  startTime: string;
  endTime: string;
  capacity: number;
}

export interface SlotUpdate {
  slotDate?: string;
  startTime?: string;
  endTime?: string;
  capacity?: number;
  status?: string;
}

export interface Booking {
  id: number;
  slotId: number;
  offerId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  peopleCount: number;
  /** @nullable */
  specialNote?: string | null;
  bookingReference: string;
  status: string;
  createdAt: string;
  slot?: Slot;
  offer?: Offer;
}

export interface BookingInput {
  slotId: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  /** @minimum 1 */
  peopleCount: number;
  specialNote?: string;
}

export interface PublicOffer {
  id: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  discountPercent: number;
  status: string;
  /** @nullable */
  bannerUrl?: string | null;
  business: Business;
  totalCapacity: number;
  totalBooked: number;
  occupancyPercent: number;
  demandLevel: string;
  availableSlots: number;
  recentBookings?: number;
  isTrending?: boolean;
  isFillingFast?: boolean;
  /** @nullable */
  recommendedSlotTime?: string | null;
  endDate?: string;
  endTime?: string;
}

export interface DemandInfo {
  recentBookings: number;
  occupancyPercent: number;
  demandLevel: string;
  isFillingFast: boolean;
  isTrending: boolean;
  /** @nullable */
  recommendedSlotId: number | null;
  /** @nullable */
  recommendedSlotTime: string | null;
}

export interface PublicOfferDetail {
  id: number;
  title: string;
  description: string;
  category: string;
  originalPrice: number;
  offerPrice: number;
  discountPercent: number;
  startDate?: string;
  endDate?: string;
  startTime?: string;
  endTime?: string;
  /** @nullable */
  terms?: string | null;
  status: string;
  /** @nullable */
  bannerUrl?: string | null;
  business: Business;
  slots: Slot[];
  demandInfo: DemandInfo;
}

export interface FeaturedOffersResponse {
  featured: PublicOffer[];
  trending: PublicOffer[];
  fillingFast: PublicOffer[];
}

export interface PulseBooking {
  id: number;
  offerTitle: string;
  businessName: string;
  businessType: string;
  customerName: string;
  createdAt: string;
  slotTime: string;
}

export interface DemandPulse {
  recentBookings: PulseBooking[];
  activeOffers: number;
  liveOccupancy: number;
  bookingVelocity: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardStats {
  totalOffers: number;
  activeOffers: number;
  totalBookings: number;
  totalRevenue: number;
  occupancyRate: number;
  capacityUtilization: number;
  todayBookings: number;
  pendingSlots: number;
  bookingsByStatus?: StatusCount[];
  offersByCategory?: CategoryCount[];
}

export interface ActivityItem {
  id: number;
  type: string;
  message: string;
  offerTitle: string;
  businessName: string;
  customerName: string;
  createdAt: string;
  bookingReference?: string;
}

export interface HourlyCount {
  hour: number;
  count: number;
}

export interface BusinessTypeCount {
  businessType: string;
  bookings: number;
  offers: number;
}

export interface DemandAnalytics {
  hourlyBookings: HourlyCount[];
  businessTypeBreakdown: BusinessTypeCount[];
  peakHour: number;
  conversionRate: number;
}

export interface TopOffer {
  id: number;
  title: string;
  businessName: string;
  category: string;
  totalBookings: number;
  occupancyPercent: number;
  revenue: number;
}

export type ListOffersParams = {
status?: string;
businessId?: number;
};

export type ListBookingsParams = {
offerId?: number;
slotId?: number;
};

export type ListPublicOffersParams = {
search?: string;
category?: string;
city?: string;
};

export type GetDashboardActivityParams = {
limit?: number;
};