# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartReserve — Smart Demand-Aware Offer Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # BUSINESSES
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a business
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # OFFERS
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: offerId
          in: query
          schema:
            type: integer
        - name: slotId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slot full or already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PUBLIC PORTAL
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string

    # BUSINESS
    Business:
      type: object
      required: [id, name, businessType, ownerName, phone, email, address, city, openingTime, closingTime]
      properties:
        id:
          type: integer
        name:
          type: string
        businessType:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BusinessInput:
      type: object
      required: [name, businessType, ownerName, phone, email, address, city, openingTime, closingTime]
      properties:
        name:
          type: string
        businessType:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        businessType:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string

    # OFFER
    Offer:
      type: object
      required: [id, businessId, title, description, category, originalPrice, offerPrice, discountPercent, startDate, endDate, startTime, endTime, status]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        originalPrice:
          type: number
        offerPrice:
          type: number
        discountPercent:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        terms:
          type: ["string", "null"]
        status:
          type: string
        bannerUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Slot:
      type: object
      required: [id, offerId, slotDate, startTime, endTime, capacity, bookedCount, availableCount, status]
      properties:
        id:
          type: integer
        offerId:
          type: integer
        slotDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        bookedCount:
          type: integer
        availableCount:
          type: integer
        status:
          type: string
        occupancyPercent:
          type: number
        demandLevel:
          type: string
...
    Booking:
      type: object
      required: [id, slotId, offerId, customerName, customerPhone, peopleCount, bookingReference, status, createdAt]
      properties:
        id:
          type: integer
        slotId:
          type: integer
        offerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        peopleCount:
          type: integer
        specialNote:
          type: ["string", "null"]
        bookingReference:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        slot:
          $ref: "#/components/schemas/Slot"
        offer:
          $ref: "#/components/schemas/Offer"
    BookingInput:
      type: object
      required: [slotId, customerName, customerPhone, peopleCount]
      properties:
        slotId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        peopleCount:
          type: integer
          minimum: 1
        specialNote:
          type: string

    # PUBLIC
...