# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartReserve — Smart Demand-Aware Offer Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Offer } from './offer';
import type { Slot } from './slot';

export interface Booking {
  id: number;
  slotId: number;
  offerId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  peopleCount: number;
  /** @nullable */
  specialNote?: string | null;
  bookingReference: string;
  status: string;
  createdAt: Date;
  slot?: Slot;
  offer?: Offer;
}