# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartReserve — Smart Demand-Aware Offer Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  slotId: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  /** @minimum 1 */
  peopleCount: number;
  specialNote?: string;
}