# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartReserve — Smart Demand-Aware Offer Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  name: string;
  businessType: string;
  ownerName: string;
  phone: string;
  email: string;
  address: string;
  city: string;
  openingTime: string;
  closingTime: string;
  logoUrl?: string;
  bannerUrl?: string;
}